+++ src/node/node_suspend_test.go
@@ -4,6 +4,8 @@ import (
 	"os"
 	"testing"
 	"time"
+
+	_state "github.com/mosaicnetworks/babble/src/node/state"
 )
 
 func TestAutoSuspend(t *testing.T) {
@@ -38,11 +40,11 @@ func TestAutoSuspend(t *testing.T) {
 	submitTransaction(nodes[0], []byte("the tx that will never be committed"))
 	waitSuspend(nodes, 10*time.Second, t)
 
-	if s := nodes[0].getState(); s != Suspended {
+	if s := nodes[0].GetState(); s != _state.Suspended {
 		t.Fatalf("nodes[0] should be Suspended, not %v. UndeterminedEvents: %d",
 			s, len(nodes[0].core.GetUndeterminedEvents()))
 	}
-	if s := nodes[1].getState(); s != Suspended {
+	if s := nodes[1].GetState(); s != _state.Suspended {
 		t.Fatalf("nodes[1] should be Suspended, not %v. UndeterminedEvents: %d",
 			s, len(nodes[1].core.GetUndeterminedEvents()))
 	}
@@ -64,11 +66,11 @@ func TestAutoSuspend(t *testing.T) {
 	submitTransaction(nodes[0], []byte("the tx that will never be committed"))
 	waitSuspend(nodes, 10*time.Second, t)
 
-	if s := nodes[0].getState(); s != Suspended {
+	if s := nodes[0].GetState(); s != _state.Suspended {
 		t.Fatalf("nodes[0] should be Suspended, not %v. UndeterminedEvents: %d",
 			s, len(nodes[0].core.GetUndeterminedEvents()))
 	}
-	if s := nodes[1].getState(); s != Suspended {
+	if s := nodes[1].GetState(); s != _state.Suspended {
 		t.Fatalf("nodes[1] should be Suspended, not %v. UndeterminedEvents: %d",
 			s, len(nodes[1].core.GetUndeterminedEvents()))
 	}
@@ -98,7 +100,7 @@ func waitSuspend(nodes []*Node, timeout time.Duration, t *testing.T) {
 		time.Sleep(10 * time.Millisecond)
 		done := true
 		for _, n := range nodes {
-			if n.getState() != Suspended {
+			if n.GetState() != _state.Suspended {
 				done = false
 				break
 			}