+++ CHANGELOG.md
@@ -3,8 +3,16 @@
 ## Unreleased
 
 SECURITY:
+
 FEATURES:
+
+- proxy: BREAKING API CHANGE - New OnStateChanged method in AppProxy
+
 IMPROVEMENTS:
+
+- babble: Dynamic SuspendLimit. SuspendLimit is now multiplied to the number of
+          validators.
+
 BUG FIXES:
 
 ## v0.6.2 (February 7, 2020)